+++ eo-archetype/src/main/resources/archetype-resources/src/main/eo/hello.eo
@@ -1,4 +1,4 @@
-+alias stdout org.eolang.io.Stdout
++alias stdout org.eolang.io.stdout
 
 [] > hello
   [] > say
