+++ eo-compiler/src/test/resources/org/eolang/compiler/packs/catches-alias-duplicates.yaml
@@ -3,11 +3,11 @@ xsls:
 tests:
   - /program/errors[count(error)=1]
   - /program/errors/error[@line='2']
-  - /program/metas/meta[head='alias' and tail='stdout org.eolang.io.Stdout']
+  - /program/metas/meta[head='alias' and tail='stdout org.eolang.io.stdout']
 eo: |
-  +alias stdout org.eolang.io.Stdout
-  +alias stdout org.eolang.io.Stdout
-  +alias stding org.eolang.io.Stdin
+  +alias stdout org.eolang.io.stdout
+  +alias stdout org.eolang.io.stdout
+  +alias stding org.eolang.io.stdin
 
   [] > main
     (stdout "Hello, world!").print
