+++ eo-compiler/src/test/resources/org/eolang/compiler/packs/leap-year.yaml
@@ -20,9 +20,9 @@ eo: |
   #
   # License is MIT
 
-  +alias stdout org.eolang.io.Stdout
-  +alias stdin org.eolang.io.Stdin
-  +alias scanner org.eolang.txt.Scanner
+  +alias stdout org.eolang.io.stdout
+  +alias stdin org.eolang.io.stdin
+  +alias scanner org.eolang.txt.scanner
 
   [args] > main
     [y] > leap
