+++ eo-maven-plugin/src/it/simple/src/main/eo/pixel.eo
@@ -1,4 +1,4 @@
-+alias stdout org.eolang.io.Stdout
++alias stdout org.eolang.io.stdout
 
 [] > pixel
   [] > hello
