+++ eo-maven-plugin/src/it/simple/verify.groovy
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-assert new File(basedir, 'target/generated-sources/eo/pixel.java').exists()
-assert new File(basedir, 'target/generated-sources/eo/pixel$hello.java').exists()
-assert new File(basedir, 'target/classes/pixel.class').exists()
-assert new File(basedir, 'target/classes/pixel$hello.class').exists()
+assert new File(basedir, 'target/generated-sources/eo/EOpixel.java').exists()
+assert new File(basedir, 'target/generated-sources/eo/EOpixel$hello.java').exists()
+assert new File(basedir, 'target/classes/EOpixel.class').exists()
+assert new File(basedir, 'target/classes/EOpixel$hello.class').exists()
