+++ eo-maven-plugin/src/test/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,1 @@
+EO-Version: ${project.version}
