+++ eo-maven-plugin/src/test/resources/log4j.properties
@@ -1,0 +1,7 @@
+log4j.rootLogger=WARN, CONSOLE
+
+log4j.appender.CONSOLE=org.apache.log4j.ConsoleAppender
+log4j.appender.CONSOLE.layout=com.jcabi.log.MulticolorLayout
+log4j.appender.CONSOLE.layout.ConversionPattern=[%color{%p}] %c: %m%n
+
+log4j.logger.org.eolang=DEBUG
