+++ eo-objects/src/main/eo/org/eolang/Sub.eo
@@ -1,7 +1,6 @@
 # Calculates algebraic difference between x and y
 
 +package org.eolang
-+export sub Sub
 
 [x y] > sub
   add
