+++ eo-runtime/src/main/java/org/eolang/EOif.java
@@ -24,16 +24,15 @@
 
 package org.eolang;
 
-import org.cactoos.iterable.Sorted;
 import org.eolang.sys.Args;
 import org.eolang.sys.Phi;
 
 /**
- * AND.
+ * IF.
  *
- * @since 0.1
+ * @since 0.2
  */
-public final class And implements Phi {
+public final class EOif implements Phi {
 
     /**
      * Args.
@@ -44,21 +43,17 @@ public final class And implements Phi {
      * Ctor.
      * @param arg Args
      */
-    public And(final Args arg) {
+    public EOif(final Args arg) {
         this.args = arg;
     }
 
     @Override
     public Object call() throws Exception {
-        boolean result = true;
-        for (final String key : new Sorted<>(this.args.keys())) {
-            if (key.charAt(0) != '0') {
-                continue;
-            }
-            result &= this.args.call(key, Boolean.class).equals(Boolean.TRUE);
-            if (!result) {
-                break;
-            }
+        Object result = false;
+        if (this.args.call("01", Boolean.class).equals(Boolean.TRUE)) {
+            result = this.args.get("02");
+        } else if (this.args.has("03")) {
+            result = this.args.get("03");
         }
         return result;
     }
