+++ eo-runtime/src/test/java/org/eolang/EOmulTest.java
@@ -23,29 +23,46 @@
  */
 package org.eolang;
 
+import org.eolang.sys.ArgsException;
 import org.eolang.sys.ArgsOf;
 import org.eolang.sys.Entry;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 /**
- * Test case for {@link Less}.
+ * Test case for {@link AddOf}.
  *
  * @since 0.2
  */
-public final class LessTest {
+public final class EOmulTest {
 
     @Test
-    public void compares() throws Exception {
+    public void mulsTwo() throws Exception {
         MatcherAssert.assertThat(
-            new Less(
-                new ArgsOf(
-                    new Entry("01", -1L),
-                    new Entry("02", 1L)
-                )
+            new EOmul(
+                new ArgsOf(new Entry("01", 1L), new Entry("02", -1L))
             ).call(),
-            Matchers.equalTo(true)
+            Matchers.equalTo(-1L)
+        );
+    }
+
+    @Test
+    public void mulsOneArg() throws Exception {
+        MatcherAssert.assertThat(
+            new EOmul(
+                new ArgsOf(new Entry("01", 1L))
+            ).call(),
+            Matchers.equalTo(1L)
+        );
+    }
+
+    @Test
+    public void mulsNoArgs() {
+        Assertions.assertThrows(
+            ArgsException.class,
+            () -> new EOmul(new ArgsOf()).call()
         );
     }
